requirements.txt
altgraph==0.17.4
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
fonttools==4.56.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
kiwisolver==1.4.7
macholib==1.16.3
matplotlib==3.9.4
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pymongo==4.11
pynput==1.7.7
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
tzdata==2025.1
zipp==3.21.0
